[SW-8] module refactor
