+++ Projects/App/Sources/DI/DataAssembly.swift
@@ -12,8 +12,13 @@ import Data
 
 public struct DataAssembly: Assembly {
     public func assemble(container: Container) {
-        container.register(WeatherRepository.self) { _ in
-            return DefaultWeatherRepository()
+        container.register(WeatherDataSource.self) { _ in
+            return DefaultWeatherDataSource()
+        }
+        
+        container.register(WeatherRepository.self) { resolver in
+            let dataSource = resolver.resolve(WeatherDataSource.self)!
+            return DefaultWeatherRepository(dataSource: dataSource)
         }
     }
 }
