--- Projects/Entity/Project.swift
@@ -1,16 +1,0 @@
-//
-//  Project.swift
-//  ProjectDescriptionHelpers
-//
-//  Created by 김상우 on 2023/03/31.
-//
-
-import ProjectDescription
-import ProjectDescriptionHelpers
-
-let project = Project.makeModule(
-    name: "Entity",
-    product: .staticFramework,
-    dependencies: [],
-    resources: []
-)
