--- Projects/Entity/Tests/gitkeep.swift
@@ -1,8 +1,0 @@
-//
-//  gitkeep.swift
-//  ProjectDescriptionHelpers
-//
-//  Created by 김상우 on 2023/03/31.
-//
-
-import Foundation
